Autonomous - TensorFlow Success
We managed to imlpement the tensorflow scan in our own class and to also use it with an autonomous class. Yay ^_^

We do this by saving the position in a variable called "gold" in the AutonomousTest class. However, we test for a temporary variable that always becomes -1 when the scanning stops, thus saving the true position in "gold" variable.