Movement Rotation Constants
After many calculations and measurements, we've concluded that the specific constants for the robot are approximatively 24.63 for ticks/cm (how many motor ticks needed for a cm of movement) and respectively 34.85 cm for the radius (distance from the center of the robot to a wheel).

The rotation works better now.